+++ src/coroutine/Task.cpp
@@ -32,7 +32,7 @@ auto coContext::Task::operator=(Task &&other) noexcept -> Task & {
 
 coContext::Task::~Task() { this->destroy(); }
 
-auto coContext::Task::getHash() const noexcept -> std::uint64_t {
+auto coContext::Task::getHash() const noexcept -> std::size_t {
     return std::hash<std::coroutine_handle<promise_type>>{}(this->handle);
 }
